refactor: extract tab stream controller factory
