+++ src/features/chat/AGENTS.md
@@ -19,6 +19,7 @@
 | `TabSession` | Authoritative per-tab identity, conversation binding, provider binding, lifecycle value, execution-coordinator attachment, active-turn reference, and background-work sequencing |
 | `TabRuntimeFactory` | Allocation of the provider-neutral runtime shell: per-tab DOM, `ChatState`, `TabSession`, placeholder controllers/services/UI slots, and initial draft/provider resolution |
 | `TabControllerFactory` | Installation of the renderer and selection controllers for a tab; execution controllers remain in the tab orchestration flow until their dependencies are separated |
+| `TabStreamControllerFactory` | Assembly of one tab's stream controller, provider-owned subagent history recovery, and background-work/persistence adapters |
 | `TabRuntimeCleanup` | Reverse-order, idempotent teardown of tab-owned resources; cleanup failures are collected so later resources still release |
 | `TabModelSelectionCoordinator` | Per-tab model-selection request ordering, blank-tab provider-transition serialization, and stable-draft rollback |
 | `ChatExecutionCoordinator` | One tab's provider-session binding, active execution, interaction fencing, cancellation, and disposal |
